Extract building atribute values vector from products search
