+++ saleor/product/search.py
@@ -1,16 +1,15 @@
 from collections import defaultdict
 from collections.abc import Iterable
-from typing import TYPE_CHECKING, Union
+from typing import TYPE_CHECKING
 
 from django.conf import settings
 from django.contrib.postgres.search import SearchQuery, SearchRank
 from django.db.models import F, Q, Value, prefetch_related_objects
 
-from ..attribute import AttributeInputType
-from ..attribute.models import AssignedProductAttributeValue, Attribute, AttributeValue
+from ..attribute.models import AssignedProductAttributeValue, AttributeValue
+from ..attribute.search import get_search_vectors_for_attribute_values
 from ..core.postgres import FlatConcatSearchVector, NoValidationSearchVector
 from ..core.utils.batches import queryset_in_batches
-from ..core.utils.editorjs import clean_editor_js
 from ..page.models import Page
 from ..product.models import Product
 
@@ -148,8 +147,8 @@ def generate_attributes_search_vector_value(
             : settings.PRODUCT_MAX_INDEXED_ATTRIBUTE_VALUES
         ]
 
-        search_vectors += get_search_vectors_for_values(
-            attribute, values, page_id_to_title_map=page_id_to_title_map
+        search_vectors += get_search_vectors_for_attribute_values(
+            attribute, values, page_id_to_title_map=page_id_to_title_map, weight="B"
         )
     return search_vectors
 
@@ -168,92 +167,10 @@ def generate_attributes_search_vector_value_with_assignment(
         values = assigned_attribute.values.all()[
             : settings.PRODUCT_MAX_INDEXED_ATTRIBUTE_VALUES
         ]
-        search_vectors += get_search_vectors_for_values(attribute, values)
-    return search_vectors
-
-
-def get_search_vectors_for_values(
-    attribute: Attribute,
-    values: Union[list, "QuerySet"],
-    page_id_to_title_map: dict[int, str] | None = None,
-) -> list[NoValidationSearchVector]:
-    search_vectors = []
-
-    input_type = attribute.input_type
-    if input_type in [AttributeInputType.DROPDOWN, AttributeInputType.MULTISELECT]:
-        search_vectors += [
-            NoValidationSearchVector(Value(value.name), config="simple", weight="B")
-            for value in values
-        ]
-    elif input_type == AttributeInputType.RICH_TEXT:
-        search_vectors += [
-            NoValidationSearchVector(
-                Value(clean_editor_js(value.rich_text, to_string=True)),
-                config="simple",
-                weight="B",
-            )
-            for value in values
-        ]
-    elif input_type == AttributeInputType.PLAIN_TEXT:
-        search_vectors += [
-            NoValidationSearchVector(
-                Value(value.plain_text), config="simple", weight="B"
-            )
-            for value in values
-        ]
-    elif input_type == AttributeInputType.NUMERIC:
-        unit = attribute.unit
-        search_vectors += [
-            NoValidationSearchVector(
-                Value(value.name + " " + unit if unit else value.name),
-                config="simple",
-                weight="B",
-            )
-            for value in values
-        ]
-    elif input_type in [AttributeInputType.DATE, AttributeInputType.DATE_TIME]:
-        search_vectors += [
-            NoValidationSearchVector(
-                Value(value.date_time.strftime("%Y-%m-%d %H:%M:%S")),
-                config="simple",
-                weight="B",
-            )
-            for value in values
-        ]
-    elif input_type in [
-        AttributeInputType.REFERENCE,
-        AttributeInputType.SINGLE_REFERENCE,
-    ]:
-        # for now only AttributeEntityType.PAGE is supported
-        search_vectors += [
-            NoValidationSearchVector(
-                Value(
-                    get_reference_attribute_search_value(
-                        value, page_id_to_title_map=page_id_to_title_map
-                    )
-                ),
-                config="simple",
-                weight="B",
-            )
-            for value in values
-            if value.reference_page_id is not None
-        ]
-    return search_vectors
-
-
-def get_reference_attribute_search_value(
-    attribute_value: AttributeValue, page_id_to_title_map: dict[int, str] | None = None
-) -> str:
-    """Get search value for reference attribute."""
-    if attribute_value.reference_page_id:
-        if page_id_to_title_map:
-            return page_id_to_title_map.get(attribute_value.reference_page_id, "")
-        return (
-            attribute_value.reference_page.title
-            if attribute_value.reference_page
-            else ""
+        search_vectors += get_search_vectors_for_attribute_values(
+            attribute, values, weight="B"
         )
-    return ""
+    return search_vectors
 
 
 def search_products(qs, value):
